suricata-oinkmaster-updater: dont exit with error if missing requirements
According to Debian Policy 9.5 (cronjobs):

<< [scripts] should check if all necessary programs are installed before they
try to execute them. Otherwise, problems will arise when a package was removed
but not purged since configuration files are kept on the system in this
situation.>>

So, this script can be called without oinkmaster and suricatasc installed.
If that's the case, refuse to run instead of the previous hard error.

Closes: #834029
Signed-off-by: Arturo Borrero Gonzalez <arturo.borrero.glez@gmail.com>