suricata-oinkmaster: move the update script to /usr/sbin
And put a dummy cronjob to call it.
While at it, configure oinkmaster to use_external_bins = 0,
so it doesn't require wget to work.

We install to /usr/sbin instead of /usr/bin because oinkmaster
requires it to be run as root, since it writes to /etc.

This is something to review in the future. I would like to put
rulesets somewhere under /var instead of /etc.

Closes: #838129
Signed-off-by: Arturo Borrero Gonzalez <arturo.borrero.glez@gmail.com>