suricata: migrate from old split binary scheme
We had 2 binary packages: suricata and suricata-hyperscan
Both installed the suricata binary, one with hyperscan link.
We then diverted the binary. This is the old pattern.

The new pattern is using only a 'suricata' binary package which
detects at runtime if we can go with hyperscan.

The last package using the split approach was 3.2-2.

Gracefully migrate from the old split binary scheme.

This patch is a joint work between Sascha and Arturo.

Closes: #855573
Signed-off-by: Sascha Steinbiss <satta@debian.org>
Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>