suricata: remove Build-Conflict with libhtp-dev
This conflict was meant to prevent suricata to be built with the old libhtp
package, which is in debian no longer. Instead, I'm creating a new src:libhtp
package and this src:suricata package should indeed be built using this new
one.

While at it,  make explicit in d/rules that the external libhtp package should
be used.

Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>