suricata: support for internal unittest in autopktest
Add basic support to run the internal suricata testsuite from
debian autopkgtest.

The suricata software requires a special configure-time flag to enable
unittest/ci checks. Care should be taken to not enable this for a standard
build, since it will surely hurt performance of suricata.

In order to enable this flag, we check if we are running from
adt/debci/autpkgtest or something like this. If that's the case, then we
enable the configure-time flag.

Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>