d/changelog: add missing entry for 4.0.0-beta1-1~exp1
The 4.0.0-beta1-1~exp1 upload was inteded for debian experimental. But I made
a mistake and used 'unstable' as distribution. The upload finally landed in
unstable/testing.
But given the development was done in the experimental git branch, the
changelog entry was not present in the unstable (master) git branch.

The result is that further development / uploads to unstable / testing
lacked this changelog entry, which is important to have since the package
4.0.0-beta1-1~exp1 has been indeed present in unstable / testing.

A bit weird. Solve it by manually adding the entry.

Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>