d/rules: disable dh_auto_test
The suricata build system doesn't include build-time test, unless build with
--enable-unittests (which is not enabled in standard build due to the runtime
overhead introduced).

But we are indeed interested in running the unittest in Debian. What we are
doing right now is to let ci.debian.net (autopkgtest) run the internal unittest
by building the package there with --enable-unittests. But there is a small
problem: the suricata 'check' Makefile target runs the unittest (suricata -u)
at build time. These test fails because they require the geoip-database package
installed.
It seems that autopkgtest builds the package at first, not bounded to the test
itself. I declared the geoip-database in d/test/control, in the test itself.
But the autopkgtest dependency is not fullfilled when the package is build.

Let's try this approach:
 * disable the tests at build time
 * let's autopkgtest run the Test-Command: suricata -u

Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>